feature: login with google
This is a really big feature push... I can now login with google. But still having some troubles with loging out.

I have changed the memberschema in te server.js, and the sign up endpoint... but i still need to change the rest of the endpoints and in the front.